Added CSS classes to fix irregularity between BS4 and CoreUI
Because both BS4, CoreUI and Log Viewer all use different script and all of them are light there are a lot of irregularities. Making it so that a lot of text is unreadable.

Currently there is no way to target Log Viewer Specifically without adding classes to the blades. Which I'm not sure is a good option